refactor(core): migrate imports to amrita_sense and base.adapter; remove deprecated init function
