+++ src/amrita_core/base/adapter.py
@@ -5,9 +5,10 @@ from collections.abc import AsyncGenerator, Iterable, Sequence
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Literal
 
+from amrita_sense.logging import logger
+
 from amrita_core.config import AmritaConfig, get_config
 
-from ..logging import logger
 from ..tools.models import ToolChoice, ToolFunctionSchema
 from ..types import EmbeddingChunk, ModelPreset, ToolCall, UniResponse
